VideoList Render + Props
We rendered the VideoList under the App parent component. VideoList needs to get passed down a list of videos from the state from the App component to the child component (VideoList). We defined a property using curly braces saying: this.state.videos which is called a prop. The data we're passing from App to VideoList is called a prop.